ui-backend/requirements.txt
Flask==2.3.0
requests==2.31.0
Pillow==10.0.0
werkzeug==2.3.0